backend/requirements.txt
fastapi
uvicorn
langgraph
transformers
torch
mysql-connector-python
pydantic